feature(age-exemple): added constructor test and fixed sc
